Fix analysis and heuristic for derefs.
Captured derefs shouldn't be indiscriminately inlined, but this was
happening due to a missing analysis. This isn't unsound or anything,
as it's what the main compiler does, but it can lead to space leaks
when an object is captured by a closure, but only certain fields are
accessed.

For now, I've kept the inlining if the field is known to be a function,
or rather always called with arguments, since it results in much better
codegen for typeclass code.