Improvements to number and int const eval
Uses specialized operators for Number Ord ops rather than going the the
Ord dictionary indirection. This means const eval for Number matches the
runtime behavior in the presence of NaN, which is undefined for the
purposes of Number Ord.

Adds overflow checks to Int operators, and bails on const eval in the
case that overflow is detected.