Improve associative operator handling
Previously, there were a couple of baroque routines for reassociating
operators during eval advantageously. It only did this to a fixed depth
because this can become quadratic. I've replaced this with a single
baroque routine that solves it completely, const evaling anything it
can, and reassociating everything to the left. To do this, I've added
a new SemAssocOp node that's similar to SemRef, which buffers the
operator spine. Then during quote we actually reassociate it.

I've also added special support for compiling concatArray calls to array
spreads, which both looks nicer and is likely faster at runtime.